Fix eqeq, reformat some of the tests
